# docs/openapi.yaml
openapi: 3.0.3
info:
  title: QR Restaurant API
  version: 1.0.0
  description: QR-based dine-in restaurant ordering backend.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /ws/sessions/{sessionToken}:
    get:
      tags: [Realtime]
      summary: WebSocket for live order updates
      parameters:
        - in: path
          name: sessionToken
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "101":
          description: Switching protocols
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a staff or owner account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          $ref: "#/components/responses/UserEnvelope"
        "400":
          $ref: "#/components/responses/ErrorEnvelope"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Log in and get a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/AuthEnvelope"
        "400":
          $ref: "#/components/responses/ErrorEnvelope"
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/UserEnvelope"
        "401":
          $ref: "#/components/responses/ErrorEnvelope"
  /api/v1/restaurants:
    get:
      tags: [Restaurants]
      summary: List restaurants owned by the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          $ref: "#/components/responses/RestaurantListEnvelope"
    post:
      tags: [Restaurants]
      summary: Create a restaurant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantRequest"
      responses:
        "201":
          $ref: "#/components/responses/RestaurantEnvelope"
  /api/v1/restaurants/{restaurantId}:
    get:
      tags: [Restaurants]
      summary: Get a restaurant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      responses:
        "200":
          $ref: "#/components/responses/RestaurantEnvelope"
    patch:
      tags: [Restaurants]
      summary: Update a restaurant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRestaurantRequest"
      responses:
        "200":
          $ref: "#/components/responses/RestaurantEnvelope"
    delete:
      tags: [Restaurants]
      summary: Delete a restaurant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantId"
      responses:
        "204":
          description: No content
...
  /api/v1/upload/image:
    post:
      tags: [Upload]
      summary: Upload a restaurant asset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                context:
                  type: string
                  enum: [logo, banner, menu_item]
      responses:
        "201":
          $ref: "#/components/responses/UploadEnvelope"
  /api/v1/public/scan:
    get:
      tags: [Public]
      summary: Start or reuse a table session from a QR token
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/ScanEnvelope"
  /api/v1/public/sessions/{sessionToken}:
    get:
      tags: [Public]
      summary: Get public session status
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          $ref: "#/components/responses/PublicSessionEnvelope"
  /api/v1/public/sessions/{sessionToken}/menu:
    get:
      tags: [Public]
      summary: Get the customer menu view
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          $ref: "#/components/responses/PublicMenuEnvelope"
  /api/v1/public/sessions/{sessionToken}/orders:
    get:
      tags: [Public]
      summary: List customer orders
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          $ref: "#/components/responses/OrderListEnvelope"
    post:
      tags: [Public]
      summary: Place a customer order
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      responses:
        "201":
          $ref: "#/components/responses/OrderEnvelope"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RestaurantId:
      in: path
      name: restaurantId
      required: true
      schema:
        type: string
        format: uuid
...
    SessionToken:
      in: path
      name: sessionToken
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorEnvelope:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UserEnvelope:
      description: Single user envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UserEnvelope"
...
    AuthEnvelope:
      description: Login envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthEnvelope"
    RestaurantEnvelope:
      description: Restaurant envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RestaurantEnvelope"
    RestaurantListEnvelope:
      description: Restaurant list envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RestaurantListEnvelope"
...
    OrderListEnvelope:
      description: Order list envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OrderListEnvelope"
    OrderEnvelope:
      description: Order envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OrderEnvelope"
...
    UploadEnvelope:
      description: Upload response envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UploadEnvelope"
    ScanEnvelope:
      description: Scan response envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ScanEnvelope"
    PublicSessionEnvelope:
      description: Public session envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PublicSessionEnvelope"
    PublicMenuEnvelope:
      description: Public menu envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PublicMenuEnvelope"
  schemas:
    ErrorBody:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          nullable: true
    Meta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
          format: int64
    BaseEnvelope:
      type: object
      properties:
        success:
          type: boolean
        request_id:
          type: string
        error:
          $ref: "#/components/schemas/ErrorBody"
        meta:
          $ref: "#/components/schemas/Meta"
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [OWNER, STAFF]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                access_token:
                  type: string
                token_type:
                  type: string
                expires_in:
                  type: integer
                user:
                  $ref: "#/components/schemas/UserResponse"
    AuthEnvelope:
      allOf:
        - $ref: "#/components/schemas/AuthResponse"
    UserEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/UserResponse"
...
    CreateRestaurantRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
    UpdateRestaurantRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        is_active:
          type: boolean
          nullable: true
    Restaurant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
        theme:
          $ref: "#/components/schemas/Theme"
    RestaurantEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Restaurant"
    RestaurantListEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Restaurant"
...
    Theme:
      type: object
      properties:
        id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        primary_color:
          type: string
        secondary_color:
          type: string
        logo_url:
          type: string
          nullable: true
        banner_url:
          type: string
          nullable: true
        font_style:
          type: string
        dark_mode:
          type: boolean
...
    ScanResponse:
      type: object
      properties:
        session_token:
          type: string
          format: uuid
        restaurant:
          type: object
          properties:
            name:
              type: string
            slug:
              type: string
        table_code:
          type: string
        theme:
          $ref: "#/components/schemas/Theme"
    ScanEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ScanResponse"
    PublicSessionResponse:
      type: object
      properties:
        session_token:
          type: string
          format: uuid
        status:
          type: string
        table_code:
          type: string
        started_at:
          type: string
    PublicSessionEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PublicSessionResponse"
    PublicMenuResponse:
      type: object
      properties:
        restaurant_name:
          type: string
        categories:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              sort_order:
                type: integer
              items:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                    name:
                      type: string
                    description:
                      type: string
                      nullable: true
                    image_url:
                      type: string
                      nullable: true
                    price_cents:
                      type: integer
                    options_config:
                      type: object
                    is_available:
                      type: boolean
                    sort_order:
                      type: integer
    PublicMenuEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PublicMenuResponse"
...
    PlaceOrderRequest:
      type: object
      required: [items]
      properties:
        idempotency_key:
          type: string
          format: uuid
          nullable: true
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [menu_item_id, quantity]
            properties:
              menu_item_id:
                type: string
                format: uuid
              quantity:
                type: integer
              selected_options:
                type: object
                additionalProperties: true
        notes:
          type: string
          nullable: true
    OrderItemResponse:
      type: object
      properties:
        name_snapshot:
          type: string
        price_cents_snapshot:
          type: integer
        quantity:
          type: integer
        selected_options:
          type: object
          additionalProperties: true
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        idempotency_key:
          type: string
          nullable: true
        status:
          type: string
        total_cents:
          type: integer
        notes:
          type: string
          nullable: true
        created_at:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
    OrderEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Order"
    OrderListEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  session_id:
                    type: string
                    format: uuid
                  table_code:
                    type: string
                  status:
                    type: string
                  total_cents:
                    type: integer
                  notes:
                    type: string
                    nullable: true
                  created_at:
                    type: string
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/OrderItemResponse"
...
    UploadEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                url:
                  type: string
                size_bytes:
                  type: integer
                mime_type:
                  type: string
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/BaseEnvelope"
        - type: object
          properties:
            success:
              type: boolean
            error:
              $ref: "#/components/schemas/ErrorBody"